% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azureBlobCall.R
\name{azureBlobCall}
\alias{azureBlobCall}
\title{Azure Blob Storage API call}
\usage{
azureBlobCall(url, verb = "GET", key, ..., simple = TRUE, tz = "GMT")
}
\arguments{
\item{url}{Azure Blob container endpoint}

\item{verb}{Name of verb that you're intending to use components with}

\item{key}{The primary or secondary access key as-is from Azure Portal}

\item{...}{Additional parameters to httr::VERB}

\item{simple}{API call response content, or supplemented}

\item{tz}{Optional parameter for setting time zone for call}
}
\value{
content The response from the API
}
\description{
This function enacts a call to the Azure Blob storage
}

