% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/azureBlobPrep.R
\name{azureBlobPrep}
\alias{azureBlobPrep}
\title{Azure Blob Storage interactions prep}
\usage{
azureBlobPrep(url, verb = "GET", key, tz = "GMT")
}
\arguments{
\item{url}{Azure Blob container endpoint}

\item{verb}{Name of verb that you're intending to use components with}

\item{key}{The primary or secondary access key as-is from Azure Portal}

\item{tz}{Optional parameter for setting time zone for call}
}
\value{
components List containing various required values / objects
}
\description{
This function produces all the various compononents
needed to use the Azure Blob API
}

