% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_analysis_site.R
\name{build_analysis_site}
\alias{build_analysis_site}
\title{Build Analysis Site}
\usage{
build_analysis_site(pkg = ".", ...)
}
\arguments{
\item{pkg}{Path to package.}

\item{...}{additional arguments passed to \code{\link[pkgdown:build_site]{pkgdown::build_site()}}}
}
\description{
\bold{\preformatted{NOT WORKING YET
}}
\code{build_analysis_site()} is a wrapper for \code{\link[pkgdown:build_site]{pkgdown::build_site()}} that adds an 'Analysis' menu
containing rendered versions of all .Rmd files in \verb{analysis/}
}
\details{
When run, \code{build_analysis_site()}:
\enumerate{
\item Reads base \link{pkgdown} settings from \verb{pkgdown/_base.yml}
\item Writes base settings to \verb{pkgdown/_pkgdown.yml}
\item Creates a template using \code{\link[pkgdown:templates]{pkgdown::template_navbar()}} and inserts an \code{analysis} menu with
links to html versions of each .Rmd file in \verb{analysis/}
\item Writes the template to \verb{pkgdown/_pkgdown.yml}
\item Runs \code{\link[pkgdown:clean_site]{pkgdown::clean_site()}} and \code{\link[pkgdown:build_site]{pkgdown::build_site()}}
\item Creates a \verb{_site.yml} file based on the final \verb{_pkgdown.yml} that clones the \link{pkgdown} navbar
in a temporary build directory
\item Copies the following from \verb{analysis/} into the build directory: \verb{*.Rmd}, \verb{data/}, \verb{assets/}
\item Changes \code{html_notebook} to \code{html_document}
\item Builds a site using \code{\link[rmarkdown:render_site]{rmarkdown::render_site()}} using a modified \code{html_document} settings to
render files with the look and feel of \code{html_notebook}
\item Moves the rendered files to \verb{docs/}: \verb{*.html}, \verb{data/}, \verb{assets/} (excluding \code{index.html})
}

\code{build_analysis_site()} will fail with an error if there are no files in \verb{analysis/*.Rmd}.
}
\examples{
\dontrun{
build_analysis_site()
}
}
