% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{imgURL}
\alias{imgURL}
\title{Produce link to file stored on github repo}
\usage{
imgURL(x, repo = "stephlocke/lazyCDN", branch = "master")
}
\arguments{
\item{x}{file name}

\item{repo}{username/repo}

\item{branch}{If in gh-pages or something}
}
\description{
Produce link to file stored on github repo
}
\examples{
imgURL("blah.png")
}

