% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stephPDFSlideStyle.R
\name{stephPDFSlideStyle}
\alias{stephPDFSlideStyle}
\title{Doc template for presentations}
\usage{
stephPDFSlideStyle(template = NULL, pandoc_args = NULL, ...)
}
\arguments{
\item{template}{Which template to use}

\item{pandoc_args}{Any additional pandoc arguments (in a list)}

\item{...}{Arguments for `rmarkdown::beamer_presentation`}
}
\description{
Doc template for presentations
}

