% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stephRevealSlideStyle.R
\name{stephRevealSlideStyle}
\alias{stephRevealSlideStyle}
\title{Doc template for presentations}
\usage{
stephRevealSlideStyle(css = NULL, options = NULL, ...)
}
\arguments{
\item{css}{Stylesheet}

\item{options}{Options to pass Reveal}

\item{...}{Other parameters for}
}
\description{
Doc template for presentations
}

