% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stephSlideStyle.R
\name{stephSlideStyle}
\alias{stephSlideStyle}
\title{Doc template for presentations}
\usage{
stephSlideStyle(css = NULL, ...)
}
\arguments{
\item{css}{Stylesheet}

\item{...}{Other parameters for}
}
\description{
Doc template for presentations
}

