% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunc.R
\name{myfunc}
\alias{myfunc}
\title{Function designed purely to show unit testing}
\usage{
myfunc(a = 1, b = 2, c = "blah")
}
\arguments{
\item{a}{numeric}

\item{b}{numeric}

\item{c}{character}
}
\description{
Function designed purely to show unit testing
}

