% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{rtraining-package}
\alias{rtraining}
\alias{rtraining-package}
\title{rtraining: R Training}
\description{
My notes and experiences learning R and RStudio, bundled as an R package (work-in-progress), and published to GitHub Pages using Quarto.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jabenninghoff.github.io/rtraining/}
  \item \url{https://github.com/jabenninghoff/rtraining}
  \item Report bugs at \url{https://github.com/jabenninghoff/rtraining/issues}
}

}
\author{
\strong{Maintainer}: John Benninghoff \email{jbenninghoff@mac.com} (\href{https://orcid.org/0000-0002-6230-4742}{ORCID})

}
\keyword{internal}
