% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtraining-package.R
\docType{package}
\name{rtraining-package}
\alias{rtraining}
\alias{rtraining-package}
\title{rtraining: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@oliverwyman.com}

}
\keyword{internal}
