% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtraining-package.R
\docType{package}
\name{rtraining-package}
\alias{rtraining}
\alias{rtraining-package}
\title{rtraining: R Training Resources, Guides, Tips, and Knowledge Base}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Houses variouse material realted to teaching R.
}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@oliverwyman.com}

}
\keyword{internal}
